backend/requirements.txt
mysql-connector-python==8.2.0
pymongo==4.6.1
redis==5.0.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
fastapi
uvicorn
pydantic
